+++ src/main/CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 	- renamed all BlockDirector associated files
 
 ### Fixed
+- BlockVerticalRepeater not powering properly
+- BlockVerticalRepeater not notifying neighbors properly
 - BlockGate not transitioning to proper state on activated
 - BlockDirector displaying particles while off